add data validation to Owner, Pet, and Vaccine DTOs and controllers

+++ api/Controllers/OwnerController.cs
@@ -19,15 +19,21 @@ namespace api.Controllers
         [Route("owners")]
         public async Task<IActionResult> GetAllOwners()
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var owners = await _ownerRepo.GetAllOwners();
             var ownerDto = owners.Select(o => o.ToOwnerDto());
             return Ok(ownerDto);
         }
 
         [HttpGet]
-        [Route("owner/{id}")]
+        [Route("owner/{id:int}")]
         public async Task<IActionResult> GetOwnerById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var owner = await _ownerRepo.GetOwnerById(id);
 
             if (owner == null)
@@ -42,15 +48,21 @@ namespace api.Controllers
         [Route("owner")]
         public async Task<IActionResult> CreateOwner([FromBody] CreateOwnerRequestDto ownerDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var ownerModel = ownerDto.ToOwnerFromCreateDto();
             await _ownerRepo.CreateOwner(ownerModel);
             return CreatedAtAction(nameof(GetOwnerById), new { id = ownerModel.Id }, ownerModel.ToOwnerDto());
         }
 
         [HttpPut]
-        [Route("owner/{id}")]
+        [Route("owner/{id:int}")]
         public async Task<IActionResult> UpdateOwner([FromRoute] int id, [FromBody] UpdateOwnerRequestDto updateDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var ownerModel = await _ownerRepo.UpdateOwner(id, updateDto);
 
             if (ownerModel == null)
@@ -63,9 +75,12 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route(("owner/{id}"))]
+        [Route(("owner/{id:int}"))]
         public async Task<IActionResult> DeleteOwner([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var ownerModel = await _ownerRepo.DeleteOwner(id);
 
             if (ownerModel == null)
